requirements.txt
Flask>=3.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
requests>=2.31.0
playwright>=1.40.0
Authlib>=1.3.0
pytz>=2024.1
selenium>=4.15.0
google-cloud-storage>=2.10.0

